requirements.txt
torch
transformers
accelerate
datasets
tensorboard
pyyaml
tqdm
safetensors
pandas
matplotlib
